requirements.txt
Django==1.4
PIL==1.1.7
django-imagekit==2.0.1
django-tastypie==0.9.11
mimeparse==0.1.3
python-dateutil==1.5
six==1.1.0
wsgiref==0.1.2
yolk==0.4.3
